t2v_metrics/models/vqascore_models/perceptionlm/requirements.txt
numpy==2.1.2
omegaconf==2.3.0
msgspec==0.19.0
rouge-score==0.1.2
sacrebleu==2.5.1
sentencepiece==0.2.0
tiktoken==0.9.0
blobfile==3.0.0
wandb==0.19.8
viztracer==1.0.3
lm-eval==0.4.8
scipy==1.15.2
pynvml==12.0.0
orjson==3.10.15
einops==0.8.1
pillow==11.0.0
pyahocorasick==2.1.0
iopath==0.1.10
torchdata==0.11.0
torchcodec==0.1.0
timm==1.0.15
decord==0.6.0
opencv-python==4.11.0.86
pycocoevalcap==1.2
scikit-learn==1.6.1
scipy==1.15.2
sentencepiece==0.2.0
tokenizers==0.21.1
webdataset==0.2.111
fsspec
datatrove
ftfy
